requirements/base.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-environ>=0.11.0
psycopg2-binary>=2.9.9
drf-spectacular>=0.27.0
django-cors-headers>=4.3.0
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0
django-ratelimit>=4.1.0
pytest>=7.4.0
pytest-django>=4.5.2
factory-boy>=3.3.0
faker>=20.0.0
gunicorn>=21.2.0
whitenoise>=6.6.0
sentry-sdk>=1.40.0
django_cors_headers